Move and rename NotificationEmail to dto package
