+++ src/main/java/pl/edu/agh/racing/cmodel/dto/NotificationEmailDto.java
@@ -1,4 +1,4 @@
-package pl.edu.agh.racing.cmodel.model;
+package pl.edu.agh.racing.cmodel.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class NotificationEmail {
+public class NotificationEmailDto {
     private String subject;
     private String recipient;
     private String body;
